requirements.txt
Django==1.8.3
argparse==1.2.1
bitcoin==1.1.35
bottle==0.12.8
django-filter==0.10.0
djangorestframework==3.2.0
docopt==0.6.2
ethereum-serpent==2.0.0
leveldb==0.193
miniupnpc==1.9
psycopg2==2.6.1
py==1.4.30
pyethash==23.1
pyethereum==0.9.2
pysha3==0.3
pytest==2.7.2
pytest-timeout==0.4
pytz==2015.4
repoze.lru==0.6
requests==2.7.0
rlp==0.3.8
structlog==15.2.0
waitress==0.8.9
wsgiref==0.1.2
